core: split SITEINFOs into separate SITEINFO.js
